settings: alembic 코드 수정
- 스키마 추가
- 테이블 존재하는 경우만 삭제 진행

+++ migration/versions/20241121_1436_20a7debd99c9_add_cafe24_order_and_payment_table.py
@@ -47,6 +47,7 @@ def upgrade():
                 server_default=sa.func.now(),
                 onupdate=sa.func.now(),
             ),
+            schema="aivelabs_sv",
         )
 
     if "cafe24_payments" in inspector.get_table_names(schema="aivelabs_sv"):
@@ -77,9 +78,17 @@ def upgrade():
                 server_default=sa.func.now(),
                 onupdate=sa.func.now(),
             ),
+            schema="aivelabs_sv",
         )
 
 
 def downgrade():
-    op.drop_table("cafe24_payments", schema="aivelabs_sv")
-    op.drop_table("cafe24_orders", schema="aivelabs_sv")
+    conn = op.get_bind()
+    inspector = reflection.Inspector.from_engine(conn)
+
+    # 테이블 존재 여부 확인
+    if "cafe24_payments" in inspector.get_table_names(schema="aivelabs_sv"):
+        op.drop_table("cafe24_payments", schema="aivelabs_sv")
+
+    if "cafe24_orders" in inspector.get_table_names(schema="aivelabs_sv"):
+        op.drop_table("cafe24_orders", schema="aivelabs_sv")
